refactor: 🔧 move `calculateNextWakeTime` to a utility module
* Extracted the `calculateNextWakeTime` function from `RecurringSnoozeView`
  and placed it in a new utility file for better code organization.
* Updated references in `RecurringSnoozeView` and `index.ts` to use the
  new utility function.
* This change improves maintainability and reusability of the wake time
  calculation logic.
